Download source only if  env is set

+++ script.sh
@@ -109,21 +109,23 @@ apt-get clean
 
 # Collect corresponding source code
 
-rm -f failed_fetch
-if [ ! -d "/vagrant" ] ; then
-  echo "FATAL - /vagrant not mounted?"
-else
-  mkdir /vagrant/source
-  cd /vagrant/source
-  dpkg --get-selections | grep -v deinstall | awk '{print $1}' | \
-    while read f ; do
+if [ -n "$DOWNLOAD_SOURCE" ] ; then
+  rm -f failed_fetch
+  if [ ! -d "/vagrant" ] ; then
+    echo "FATAL - /vagrant not mounted?"
+  else
+    mkdir /vagrant/source
+    cd /vagrant/source
+    dpkg --get-selections | grep -v deinstall | awk '{print $1}' | \
+      while read f ; do
         apt-get source $f || echo $f >failed_fetch
-    done
-fi
+      done
+    fi
 
-if [ -f "failed_fetch" ] ; then
-  echo "WARNING - some sources were not fetched:"
-  cat failed_fetch
+    if [ -f "failed_fetch" ] ; then
+      echo "WARNING - some sources were not fetched:"
+      cat failed_fetch
+    fi
 fi
 
 # Final cleanup
